+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
+### Changed
+- narrow down Item detail collections to published
+
 ### Fixed
 - indexing for published Collections